requirements.txt
asgiref==3.8.1
Django==5.2.3
django-allauth==64.2.1
dotenv==0.9.9
psycopg2-binary==2.9.10
python-dotenv==1.1.0
qrcode==8.2
Pillow==10.4.0
sqlparse==0.5.3
requests==2.31.0
PyJWT==2.8.0
cryptography>=41.0.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.3
django-allauth==64.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
google-auth==2.40.3
google-auth-oauthlib==1.2.0
google-api-python-client==2.120.0
idna==3.10
pillow==10.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.1.0
qrcode==8.2
requests==2.31.0
rsa==4.9.1
sqlparse==0.5.3
urllib3==2.5.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.3
django-allauth==64.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
google-api-core==2.25.1
google-api-python-client==2.120.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
oauthlib==3.3.1
pillow==10.4.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.2.3
python-dotenv==1.1.0
qrcode==8.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
